created report generating functions

+++ app/controllers/mgEditProfile.php
@@ -24,25 +24,19 @@ class mgEditProfile extends controller
         $this->view("/Pages/edit-profile-details", array($alert, $data));
     }
 
-    public function passwordAvailable($pwd)
+    public function isPasswordAvailable($pwd)
     {
-        $count = 0;
         $passwords = $this->model("staffModel");
-        $result = $passwords->getAllPasswords();
+        $result = $passwords->getPasswords($_SESSION['user_id']);
         if ($result) {
             foreach ($result as $row) {
-                if (password_verify($pwd, $row)) {
-                    $count = $count + 1;
-                    break;
+                if (password_verify($pwd, $row->password)) {
+                    return true;
                 }
             }
         }
 
-        if ($count == 0) {
-            return false;
-        } else {
-            return true;
-        }
+        return false;
     }
 
     public function userChangePassword()
@@ -51,24 +45,28 @@ class mgEditProfile extends controller
             $currentPwd = $_POST["curr-password"];
             $confirmPwd = $_POST["confirm-password"];
             $newPwd = $_POST["new-password"];
-            $staffmem = $this->model("staffModel");
-            $passwd = $staffmem->getUserPassword($_SESSION['user_email']);
+            $passwd = $this->model("staffModel")->getUserPassword($_SESSION['user_email']);;
+
             if (password_verify($currentPwd, $passwd->password)) {
-                if ($newPwd == $confirmPwd) {
+
+                if ($newPwd == $confirmPwd and !$this->isPasswordAvailable($newPwd)) {
                     $userModel = $this->model("userModel");
                     $hash = password_hash($newPwd, PASSWORD_BCRYPT);
                     $userModel->resetPassword($_SESSION["user_email"], $hash);
                     echo json_encode(array("msg" => 1));
                 } else if (!($newPwd == $confirmPwd)) {
                     echo json_encode(array("msg" => "new-not-match-to-confirm"));
-                    flashMessage("There is a mismatch between new and confirm passwords");
-                } else if ($this->passwordAvailable($newPwd)) {
+                    flashMessage("There is a mismatch between new and confirm passwords", 0);
+                } else if ($this->isPasswordAvailable($newPwd)) {
                     echo json_encode(array("msg" => "already-available"));
-                    flashMessage("New Password is Already Available..");
+                    flashMessage("New Password is Already Available..", 0);
+                } else {
+                    echo json_encode(array("msg" => "error"));
+                    flashMessage("error", 0);
                 }
             } else {
                 echo json_encode(array("msg" => "invalid_pwd"));
-                flashMessage("Current Password is invalid");
+                flashMessage("Current Password is invalid", 0);
             }
         }
     }
@@ -127,7 +125,7 @@ class mgEditProfile extends controller
         if (password_verify($password, $DBPassword->password) and $otp == $_SESSION['otp'] and !$this->isTimeOut()) {
             echo json_encode(array("msg" => "ok"));
             unset($_SESSION['otp']);
-        } else{
+        } else {
             echo json_encode(array("msg" => "fail"));
             unset($_SESSION['otp']);
         }
@@ -139,19 +137,22 @@ class mgEditProfile extends controller
             $newEmail = $_POST["new_email"];
             $confirmEmail = $_POST["confirm_email"];
             $staffmem = $this->model("staffModel");
-            $isEmailExist = $staffmem->emailExist($newEmail);
+            $isEmailExist = $staffmem->emailExist($newEmail, $_SESSION['user_id']);
             if ($newEmail == $confirmEmail and !$isEmailExist) {
                 $userModel = $this->model("userModel");
                 $userModel->resetEmail($_SESSION["user_email"], $newEmail);
                 $_SESSION['user_email'] = $newEmail;
-                flashMessage("Email Updated");
+                flashMessage("Email Updated", 1);
                 redirect("/mgEditProfile");
             } else if ($isEmailExist) {
-                flashMessage("Email already exists");
+                flashMessage("Email already exists", 0);
                 redirect("/mgEditProfile");
-            }
+            } // }else{
+            //     flashMessage("kelawela");
+            //     redirect("/mgEditProfile");
+            // }
         } else {
-            flashMessage("Some Error Occured..Try Again");
+            flashMessage("Some Error Occured..Try Again", 0);
             redirect("/mgEditProfile");
         }
     }
@@ -159,15 +160,34 @@ class mgEditProfile extends controller
     public function setPersonalDetails()
     {
         isLoggedIn();
-        $staffMem = $this->model("staffModel");
-        $result = $staffMem->setPersonalInfo($_SESSION['user_id'], $_POST['fName'], $_POST['lName'], $_POST['gender'], $_POST['lane1'], $_POST['lane2'], $_POST['lane3'], $_POST['mob-no'], $_POST['mob-no2'], $_POST['image']);
+
+        $count = 0;
+        $phone = $this->model("staffModel")->getUserPhoneNumbers($_SESSION['user_id']);
+        foreach ($phone as $row) {
+            if ($row->userId) {
+                $count++;
+            }
+        }
+
+
+        $result = '';
+        if ($count == 2) {
+            if (!empty($phone[0]->phone)) {
+                $result = $this->model("staffModel")->setPersonalInfo($_SESSION['user_id'], $_POST['fName'], $_POST['lName'], $_POST['gender'], $_POST['lane1'], $_POST['lane2'], $_POST['lane3'], $_POST['mob-no2'], $_POST['image'], $phone[0]->phone, $phone[1]->phone);
+            } else {
+                $result = $this->model("staffModel")->setPersonalInfo($_SESSION['user_id'], $_POST['fName'], $_POST['lName'], $_POST['gender'], $_POST['lane1'], $_POST['lane2'], $_POST['lane3'], $_POST['mob-no2'], $_POST['image'], $phone[1]->phone, "empty");
+            }
+        } else if ($count == 1) {
+
+            $result = $this->model("staffModel")->setPersonalInfo($_SESSION['user_id'], $_POST['fName'], $_POST['lName'], $_POST['gender'], $_POST['lane1'], $_POST['lane2'], $_POST['lane3'], $_POST['mob-no2'], $_POST['image'], $phone[1]->phone);
+        }
 
         if ($result) {
             echo "done";
-            flashMessage("Personal Details Updated");
+            flashMessage("Personal Details Updated", 1);
         } else {
             echo "fail";
-            flashMessage("Personal Details not Updated");
+            flashMessage("Personal Details not Updated", 0);
         }
     }
 }
